Restructure server
