+++ mula/scheduler/server/handlers/__init__.py
@@ -1,0 +1,6 @@
+from .health import HealthAPI
+from .metrics import MetricsAPI
+from .queues import QueueAPI
+from .root import RootAPI
+from .schedulers import SchedulerAPI
+from .tasks import TaskAPI
